% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanTrackTitle}
\alias{cleanTrackTitle}
\title{Clean file names to get track titles}
\usage{
cleanTrackTitle(files)
}
\arguments{
\item{files}{Character vector of files}
}
\value{
Character vector of clean file names
}
\description{
Keeps only the basename without extensions. Handles URLs too.
}
\examples{
tracks <- c("track1.mp3",
            "files/track2.wav",
            "www.example.com/files/track3.mp3?year=2022",
            "files/track4.mp3")
cleanTrackTitle(tracks)

}
