% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{runHowlerExample}
\alias{runHowlerExample}
\alias{availableHowlerExamples}
\title{Run \code{\{howler\}} Example Applications}
\usage{
runHowlerExample(example = "basic", ...)

availableHowlerExamples()
}
\arguments{
\item{example}{Name of the example to load. Current examples include:
\describe{
\item{basic}{Basic example of \code{howler} in use}
\item{full}{Basic example of using all buttons available in \code{howler}}
\item{module}{Example of using the \code{howlerModule}}
\item{server}{Example showing server-side functionality}
}}

\item{...}{Optional arguments to send to \code{shiny::runApp}}
}
\description{
Run \code{\{howler\}} Example Applications
}
\examples{
availableHowlerExamples()

if (interactive()) {
  library(shiny)
  library(howler)

  runHowlerExample("basic")
}

}
