% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{howlerCurrentTrack}
\alias{howlerCurrentTrack}
\title{Current Track}
\usage{
howlerCurrentTrack(id)
}
\arguments{
\item{id}{ID given to the current track label. For it to work with the \code{\link{howlerPlayer}}, the ID
must match that of the \code{howlerPlayer}.}
}
\value{
A \code{div} tag that will be linked to the \code{\link{howlerPlayer}} to show the current track.
}
\description{
A way to display the current track in the UI without having to communicate with the server.
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- fluidPage(
    title = "howler.js Player",
    howler(elementId = "sound", "audio/sound.mp3"),
    howlerCurrentTrack("sound"),
    howlerPlayPauseButton("sound")
  )

  server <- function(input, output, session) {
  }

  shinyApp(ui, server)
}

}
