% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{howlerModuleUI}
\alias{howlerModuleUI}
\alias{howlerModuleServer}
\title{Howler.js Module}
\usage{
howlerModuleUI(id, files)

howlerModuleServer(input, output, session)
}
\arguments{
\item{id}{ID}

\item{files}{Vector of files}

\item{input, output, session}{Standard \code{shiny} input, output and session}
}
\description{
Howler.js Module
}
