% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-side.R
\name{howlerServer}
\alias{howlerServer}
\alias{changeTrack}
\alias{addTrack}
\alias{playHowl}
\alias{pauseHowl}
\alias{stopHowl}
\alias{seekHowl}
\title{Update howler.js Server-Side}
\usage{
changeTrack(session, id, track)

addTrack(session, id, track, play_track = FALSE)

playHowl(session, id)

pauseHowl(session, id)

stopHowl(session, id)

seekHowl(session, id, seek)
}
\arguments{
\item{session}{Shiny session}

\item{id}{ID of the \code{howlerPlayer} to update}

\item{track}{Base name of the file to change to. If the file is not included in the player nothing will happen.}

\item{play_track}{Logical, should the new track be played on addition?}

\item{seek}{Time (in seconds) to set the position of the track}
}
\value{
Updates the the state of the specified \code{howlerPlayer} in the shiny application.
}
\description{
Change the state of the howler player from the server.

\code{playHowl}, \code{pauseHowl} and \code{stopHowl} will all be applied to the current track.

\code{changeTrack} will update the track to the file specified. This file must be included when the player
is initialised, otherwise it won't change the track.

\code{addTrack} will add a new track to the specified player.
}
\examples{
if (interactive()) {
  library(shiny)

  tracks <- c("audio/track1.mp3", "audio/track2.mp3")

  ui <- fluidPage(
    title = "howler.js Player",
    useHowlerJS(),
    selectInput("track", "Select Track", basename(tracks)),
    howlerPlayer("howler", tracks),
    howlerPlayPauseButton("howler")
  )

  server <- function(input, output) {
    observeEvent(input$track, changeHowlerTrack("howler", input$track))
  }

  runShiny(ui, server)
}

}
