% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{useHowlerJS}
\alias{useHowlerJS}
\title{Enable \code{HowlerJS}}
\usage{
useHowlerJS(spatial = FALSE)
}
\arguments{
\item{spatial}{Logical, whether or not the spatial aspects of \code{howler.js} are required. Default is set to \code{FALSE}.
(Unless writing custom functionality, this is not required).}
}
\value{
HTML tags that link to relevant JavaScript and CSS files used to display and run \code{howler.js}.
}
\description{
Add \code{useHowlerJS} to the UI of a shiny application so that the necessary resources can be added to the application.
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- fluidPage(
    title = "Initial Title",
    useHowlerJS(),
    howlerPlayer("sound"),
    playPauseButton("sound"),
    volumeSlider("sound")
  )

  server <- function(input, output, session) {
  }

  shinyApp(ui, server)
}

}
