% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{howlerButton}
\alias{howlerButton}
\alias{playButton}
\alias{pauseButton}
\alias{playPauseButton}
\alias{previousButton}
\alias{nextButton}
\alias{volumeUpButton}
\alias{volumeDownButton}
\title{Audio Buttons}
\usage{
howlerButton(id, button_type = HOWLER_BUTTON_TYPES, ...)

playButton(id)

pauseButton(id)

playPauseButton(id)

previousButton(id)

nextButton(id)

volumeUpButton(id, volume_change = 0.1)

volumeDownButton(id, volume_change = 0.1)
}
\arguments{
\item{id}{ID given to the button. For it to work with the \code{\link{howlerPlayer}}, the ID
must match that of}

\item{button_type}{Type of button to create. Available buttons are in the details, default set to \code{play_pause}}

\item{...}{Attributes added to the button}

\item{volume_change}{How much to change the volume by. Default is 10\%.}
}
\value{
An HTML object containing the button.
}
\description{
Buttons that can be used to change the audio of
}
\details{
The following \code{button_type} are available to create:

\describe{
\item{\code{play_pause}}{(default) Button that switches}
}

When using a \code{play_pause} button, the icon will toggle between the play and pause button
depending on whether or not the track is playing.
}
