% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{howlerModuleUI}
\alias{howlerModuleUI}
\alias{howlerModuleServer}
\title{Howler.js Module}
\usage{
howlerModuleUI(id, files)

howlerModuleServer(input, output, session)
}
\arguments{
\item{id}{ID}

\item{files}{Vector of files}

\item{input, output, session}{Standard \code{shiny} input, output and session}
}
\value{
The server-side module will return a list of reactive objects:
\describe{
\item{playing}{Logical value whether or not the player is currently playing}
\item{track}{Name of the track currently loaded}
\item{duration}{Duration (in seconds) of the track currently loaded}
\item{seek}{Current position (in seconds) of the track currently loaded}
}
}
\description{
Howler.js Module
}
\details{
The player will have the ID \code{{id}-howler}.
}
\examples{
if (interactive()) {
  ui <- fluidPage(
    useHowlerJS(),
    howlerModuleUI("howl")
  )

  server <- function(input, output, session) {
    moduleServer("howl", howlerModuleServer)
  }
}

}
