% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-side.R
\name{changeHowlTrack}
\alias{changeHowlTrack}
\alias{playHowl}
\alias{pauseHowl}
\alias{stopHowl}
\title{Server-Side Howler}
\usage{
changeHowlTrack(session, id, file)

playHowl(session, id)

pauseHowl(session, id)

stopHowl(session, id)
}
\arguments{
\item{session}{Shiny session}

\item{id}{ID of the \code{howlerPlayer} to update}

\item{file}{File to change to}
}
\description{
This is a series of functions that can be used in the server
}
