% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{useHowlerJS}
\alias{useHowlerJS}
\title{Enable \code{HowlerJS}}
\usage{
useHowlerJS()
}
\value{
HTML tags that link to relevant JavaScript and CSS files used to display and run \code{howler.js}.
}
\description{
Add \code{useHowlerJS} to the UI of a shiny application so that the necessary resources can be added to the application.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(shiny)

ui <- fluidPage(
  title = "howler.js Player Title",
  useHowlerJS(),
  howlerPlayer("sound", "audio/track1.mp3"),
  howlerPlayPauseButton("sound"),
  howlerVolumeSlider("sound")
)

server <- function(input, output, session) {}

shinyApp(ui, server)
\dontshow{\}) # examplesIf}
}
