% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{useHowlerJS}
\alias{useHowlerJS}
\title{Enable \code{HowlerJS}}
\usage{
useHowlerJS(spatial = FALSE)
}
\value{
A script tag that enables \code{howler} to work within a shiny app.
}
\description{
Add this function to the UI of a shiny application in order for you to be able to use HowlerJS
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- fluidPage(
    title = "Initial Title",
    useHowlerJS(),
    actionButton("button", "Play Sound"),
  )

  server <- function(input, output, session) {
    observeEvent(input$button, playSound(session))
  }

  shinyApp(ui, server)
}

}
